% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apg.R
\docType{data}
\name{apg}
\alias{apg}
\title{APG families}
\format{A data frame with 1597 rows and 3 columns}
\description{
Complete synonymy of family names
}
\note{
Downloaded from \url{http://www.mobot.org/MOBOT/research/APweb/}
}
\keyword{datasets}

