% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2phylomatic.R
\name{df2phylomatic}
\alias{df2phylomatic}
\title{Phylomatic format}
\usage{
df2phylomatic(taxa, uppercase = TRUE)
}
\arguments{
\item{taxa}{A data frame with columns named family, genus, and species.}

\item{uppercase}{logical. Should the function capitalize first letters?}
}
\description{
Phylomatic format
}

