% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/noauthors.R
\name{noauthors}
\alias{noauthors}
\title{Remove the author(s) from a taxon name.}
\usage{
noauthors(taxon)
}
\arguments{
\item{taxon}{a character vector containing a single taxon}
}
\value{
a character vector
}
\description{
This attempts to remove the authorities of a taxonomic name.
}
\examples{
noauthors("Coffea arabica L.")
noauthors("Chrysophyllum argenteum subsp. nitidum (G.F.W.Meyer) T.D.Penn.")
}

