% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tpl.get.R
\name{tpl.get}
\alias{tpl.get}
\title{Get plant taxonomical and distribution data}
\usage{
tpl.get(taxa, replace.synonyms = TRUE, suggest.names = TRUE,
  suggestion.distance = 0.9, drop = c("major.group", "genus.hybrid.marker",
  "species.hybrid.marker", "nomenclatural.status.from.original.data.source",
  "ipni.id", "source.id", "publication", "collation", "page", "date"),
  apg.families = TRUE, return.synonyms = FALSE)
}
\arguments{
\item{taxa}{a character vector containing one or more taxa, without authors
see \code{\link{noauthors}} if you have a list with authorities}

\item{replace.synonyms}{should the function automatically replace synonyms?}

\item{suggest.names}{should the function try to suggest corrections for spelling errors?}

\item{suggestion.distance}{how conservative should the suggestion algorithm be?}

\item{drop}{NULL or character vector with columns from original dataset to drop.}

\item{apg.families}{Return APG families?}

\item{return.synonyms}{Return a list of synonyms instead of the regular dataset?}
}
\value{
a data frame
}
\description{
This function collects taxonomic information from The Plant List. Synonyms and
misspelled names are resolved automatically.
}
\examples{
\dontrun{
tpl.get("Myrcia lingua")
}
}

