% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{tpl}
\alias{tpl}
\alias{tpl-package}
\title{Package tpl}
\description{
Interacting with The Plant List
}
\details{
This package contains functions to gather data from The Plant List database.
Synonyms and spelling errors can be solved automatically. All data are
distributed with the package, so no internet connection is needed.
}

