% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tpl.get}
\alias{tpl.get}
\title{Get plant taxonomical and distribution data}
\usage{
tpl.get(taxa, replace.synonyms = TRUE, suggest.names = TRUE)
}
\arguments{
\item{taxa}{a character vector containing one or more taxa, without authors
see \code{\link{noauthors}} if you have a list with authorities}

\item{replace.synonyms}{should the function automatically replace synonyms?}

\item{suggest.names}{should the function try to suggest corrections for spelling errors?}
}
\value{
a data frame
}
\description{
This function collects taxonomic information from The Plant List. Synonyms and
misspelled names are resolved automatically.
}
\examples{
\dontrun{
tpl.get("Myrcia lingua")
}
}

